diagram_ai_generator/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "diagram-ai-generator"
version = "1.0.0"
description = "CLI tool for generating architecture diagrams using AI prompts"
authors = [
    {name = "Diagram AI Generator Team"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "openai>=1.0.0",
    "click>=8.0.0",
    "pydantic>=2.0.0",
    "diagrams>=0.23.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "mcp>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
    "flake8>=6.0.0",
]

[project.scripts]
diagram-ai = "diagram_ai_generator.src.application.cli.main:main"

[project.urls]
Homepage = "https://github.com/your-username/diagram-ai-generator"
Repository = "https://github.com/your-username/diagram-ai-generator"
Issues = "https://github.com/your-username/diagram-ai-generator/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["diagram_ai_generator*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true